8.16 C++ example: Salary calculation with classes
The program below uses a class, TaxTableTools, which has a tax table built in. The main function prompts for a salary, then uses a TaxTableTools function to get the tax rate. The program then calculates the tax to pay and displays the results to the user. Run the program with annual salaries of 10000, 50000, 50001, 100001 and -1 (to end the program) and note the output tax rate and tax to pay.

Modify the TaxTableTools class to use a setter function that accepts a new salary and tax rate table.
Modify the program to call the new function, and run the program again, noting the same output.
Note that the program's two classes are in separate tabs at the top.